Type adjustment

+++ librtld/include/rtld/types.hpp
@@ -2,5 +2,5 @@
 
 namespace rtld {
 typedef size_t Result;
-typedef uint32_t Handle;
+typedef uintptr_t Handle;
 }  // namespace rtld
